Add files via upload
Pipeline and BehaviorOnly are the main files to run for analysis and use the subsets. Directory dependencies should be specified here.  These updated sub-functions now rely on the data_path and results_path (where to save) that you specify in Pipeline or BehaviorOnly.

Pipeline will make a massive table of all the variables I was playing with for my analysis separated by file.  Not all of them are used.  It considers neural data.  It does not currently save the analysis results

BehaviorOnly makes a table of the behavioral analyses.  It DOES save the output in a results_path, and looks for this output before rerunning all the files.  It will add any un-processed files in the data_path and organize the output alphabetically.

Files I modified for use in my thesis.  Neural_Parser now asks for a data_path. Pipeline will make some figures.  A